[feat] 기존 view 초기화
